Multiple Heap refactor
