+++ lib/Heap.rb
@@ -1,9 +1,11 @@
 require 'Heap/version'
 
+# Main ruby-heap rb file
 module Heap
   require 'Heap/binary_heap/binary_heap'
   require 'Heap/binary_heap/binary_heap_min'
   require 'Heap/binary_heap/binary_heap_max'
+  require 'Heap/multiple_heap/multiple_heap'
   require 'Heap/multiple_heap/multiple_heap_max'
   require 'Heap/multiple_heap/multiple_heap_min'
 end
